v0.9.7
Bugs from HYV-123 resolved:
- topology.py._only_end_nodes: can work with cirular hydroobjects
- geometry.py.find_nearest_branch: can work with emtpy GeoSeries (converted to floats)

+++ hydamo_validation/geometry.py
@@ -112,7 +112,7 @@ def find_nearest_branch(branches, geometries, method="overall", maxdist=5):
                 # Since a culvert can cross a channel, it is
                 crds = geometry.geometry.coords[:]
                 dist = selectie["geometry"].apply(lambda x: max(x.distance(Point(*crds[0])),
-                                                                x.distance(Point(*crds[-1]))))
+                                                                x.distance(Point(*crds[-1])))).astype(float)
                 # dist = (
                 #     selectie.distance(Point(*crds[0]))
                 #     + selectie.distance(Point(*crds[-1]))
